# openapi-cleaned.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CLARITY Digital Twin Platform",
    "description": "Production AWS-native health data platform with comprehensive API endpoints",
    "version": "0.2.0",
    "x-logo": {
      "url": "https://clarity.novamindnyc.com/logo.png",
      "altText": "CLARITY Logo"
    },
    "contact": {
      "name": "CLARITY Support",
      "email": "support@clarity.novamindnyc.com",
      "url": "https://clarity.novamindnyc.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://clarity.novamindnyc.com/license"
    }
  },
  "paths": {
...
    "/api/v1/pat/analysis/{processing_id}": {
      "get": {
        "tags": [
          "pat-analysis",
          "pat-analysis"
        ],
        "summary": "Get PAT Analysis Results",
        "description": "Retrieve PAT (Pretrained Actigraphy Transformer) analysis results for a specific processing ID.\n\n    **Returns:**\n    - PAT model embeddings and predictions\n    - Activity pattern analysis\n    - Sleep quality metrics\n    - Circadian rhythm insights\n    - Processing metadata and status\n\n    **Status Values:**\n    - `completed`: Analysis finished successfully\n    - `processing`: Analysis still running\n    - `failed`: Analysis encountered an error\n    - `not_found`: No analysis found for this ID",
        "operationId": "get_pat_analysis_api_v1_pat_analysis__processing_id__get",
        "parameters": [
          {
            "name": "processing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Processing Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analysis results retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PATAnalysisResponse"
                }
              }
            }
          },
          "404": {
            "description": "Analysis not found"
          },
          "503": {
            "description": "Service temporarily unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/pat/models/info": {
      "get": {
        "tags": [
          "pat-analysis",
          "pat-analysis"
        ],
        "summary": "Get PAT Model Information",
        "description": "Get information about the available PAT models and current configuration",
        "operationId": "get_model_info_api_v1_pat_models_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Model Info Api V1 Pat Models Info Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PATAnalysisResponse": {
        "properties": {
          "processing_id": {
            "type": "string",
            "title": "Processing Id",
            "description": "Processing ID for the analysis"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status: completed, processing, failed, not_found"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Status message"
          },
          "analysis_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Date",
            "description": "When analysis was completed"
          },
          "pat_features": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pat Features",
            "description": "PAT model features"
          },
          "activity_embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Embedding",
            "description": "Activity embedding vector"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Additional metadata"
          }
        },
        "type": "object",
        "required": [
          "processing_id",
          "status"
        ],
        "title": "PATAnalysisResponse",
        "description": "Response model for PAT analysis results."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "details": {
            "type": "object",
            "description": "Additional error details",
            "additionalProperties": true
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token authentication"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key authentication"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://clarity-alb-1762715656.us-east-1.elb.amazonaws.com",
      "description": "Production server (AWS ALB)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
...
}